Raise better exception when wrong callback object passed
