+++ py7zr/py7zr.py
@@ -23,6 +23,7 @@
 #
 #
 """Read 7zip format archives."""
+import abc
 import collections.abc
 import contextlib
 import datetime
@@ -507,11 +508,18 @@ class SevenZipFile(contextlib.AbstractContextManager):
         return_dict: bool = False,
         callback: Optional[ExtractCallback] = None,
     ) -> Optional[Dict[str, IO[Any]]]:
-        if callback is not None and not isinstance(callback, ExtractCallback):
-            raise ValueError("Callback specified is not a subclass of py7zr.callbacks.ExtractCallback class")
-        elif callback is not None:
-            self.reporterd = Thread(target=self.reporter, args=(callback,), daemon=True)
-            self.reporterd.start()
+        try:
+            if callback is None:
+                pass
+            elif isinstance(callback, ExtractCallback):
+                self.reporterd = Thread(target=self.reporter, args=(callback,), daemon=True)
+                self.reporterd.start()
+            else:
+                raise ValueError(
+                    "Callback specified is not an instance of subclass of py7zr.callbacks.ExtractCallback class"
+                    )
+        except:
+            raise TypeError("Callback is not an instance but class, or unknown object")
         target_files: List[Tuple[pathlib.Path, Dict[str, Any]]] = []
         target_dirs: List[pathlib.Path] = []
         if path is not None:
